airflow/requirements.txt
kafka-python==2.0.2
confluent_kafka==2.1.1
airflow_provider_kafka==0.2.2